# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book해도. REST API",
    "version": "2.0.0",
    "description": "쿠키 인증, 소유권 검증, 날짜별 optimistic concurrency. 전체 일정/메모/대안 변경에 expectedRevision 필수(누락 428, 충돌 409). POST 장소 추가는 원자적 append. X-Request-ID로 오류 추적."
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "동일 출처 API"
    }
  ],
  "paths": {
...
    "/regions": {
      "get": {
        "summary": "12개 관광권과 실제 DB 장소 수",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Region"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "장소"
        ],
        "operationId": "get__regions"
      }
    },
...
    "/places": {
      "get": {
        "summary": "실제 장소 검색·지역·분류 필터",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Place"
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    },
                    "filters": {
                      "type": "object",
                      "properties": {
                        "regions": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "types": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "keyword": {
                          "type": "string"
                        }
                      }
                    },
                    "notice": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "regionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ATTRACTION",
                "RESTAURANT",
                "LODGING"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 40,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0,
              "maximum": 20000
            }
          }
        ],
        "description": "q에 지역+업종·활동+장소명 입력 가능(예: 삿포로 공방, 오타루 빈티지 옷가게). 지역 AND 의도 태그 AND 잔여 장소명으로 검색한다. 공방·공예, 공원·자연, 문화시설, 쇼핑·의류·빈티지·서점·시장, 한식·일식·양식·중식·라멘·스시·카레·카페·디저트·주점, 온천·스키·놀거리, 호텔·료칸·캠핑·호스텔 등의 표현과 동의어를 지원한다. 구체적 표현이 있으면 같은 그룹의 포괄 표현을 중복 적용하지 않는다. 입력 지역은 regionId보다 우선하며 여러 지역/유형은 그룹 내 OR로 처리한다.",
        "tags": [
          "장소"
        ],
        "operationId": "get__places"
      }
    },
    "/places/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "장소 상세 및 출처",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Place"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "장소"
        ],
        "operationId": "get__places_id_"
      }
    },
...
    "/discover": {
      "get": {
        "summary": "상황·날씨·주변·리뷰 표본 기반 추천",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Place"
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "notice": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "READY",
                        "PROVIDER_UNAVAILABLE",
                        "INSUFFICIENT_EVIDENCE",
                        "NO_CANDIDATES"
                      ]
                    },
                    "evidence": {
                      "type": "object",
                      "properties": {
                        "candidateCount": {
                          "type": "integer"
                        },
                        "inspected": {
                          "type": "integer"
                        },
                        "available": {
                          "type": "integer"
                        },
                        "failed": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 다른 사용자의 리소스",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "theme",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "weather",
                "indoor",
                "nature",
                "photo",
                "onsen",
                "food",
                "random",
                "nearby",
                "season",
                "japanese"
              ]
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "regionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tripId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "latitude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 40,
              "default": 20
            }
          }
        ],
        "description": "날씨·실내 판단은 공통 정책을 사용. 일본어 리뷰 테마는 음식점·관광지 후보 최대 10곳을 최대 2개씩 동시 조회한다(12초 총 예산). 10곳은 검사 대상의 상한이며 후보 부족·조회 실패·품질 기준 미달 시 결과가 더 적을 수 있다. 장소당 최대 5개 리뷰 표본을 확인한다. 전체 평점 3.5 이상, 전체 리뷰 5개 이상, 일본어 원문 표본 존재, 폐업 아님을 확인한다. 점수=(전체 평점×리뷰 수+3.5×20)/(리뷰 수+20). 일본어 리뷰만의 평균이나 현지인 여부를 추정하지 않는다. 리뷰 응답은 DB에 저장하지 않는다.",
        "tags": [
          "추천"
        ],
        "operationId": "get__discover"
      }
    },
    "/places/{id}/enrichment": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Google Places 한국어 실시간 상세 보강",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean"
                    },
                    "notice": {
                      "type": "string"
                    },
                    "nameKo": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "source": {
                      "type": "string"
                    },
                    "reviews": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "reviewEvidence": {
                      "type": "object"
                    },
                    "photo": {
                      "type": [
                        "object",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 다른 사용자의 리소스",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "일본어 원문 이름과 1.2km 이내 위치가 일치할 때만 연결. 사진·리뷰 출처와 번역 여부 포함. Google Place ID만 연결 테이블에 저장. 사진·리뷰 응답은 DB에 저장하지 않음.",
        "tags": [
          "장소"
        ],
        "operationId": "get__places_id_enrichment"
      }
    },
    "/place-photo": {
      "get": {
        "summary": "서명된 단기 사진 참조",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Google 사진 URL로 이동"
          },
          "401": {
            "description": "로그인 필요"
          },
          "404": {
            "description": "토큰 만료/사진 없음"
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "시스템"
        ],
        "operationId": "get__place_photo"
      }
    },
    "/places/{id}/trend": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "저장된 리뷰 추이·품질·주/월 집계",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trendType": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "recentReviewCount": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "previousReviewCount": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "growthRate": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "seasonalPatterns": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "quality": {
                      "type": "object"
                    },
                    "periods": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "scraper 실행 없음. JST 기준 완료 30일 대 이전 30일. 부분 수집·상한 도달·검증 실패·7일 초과 갱신 지연은 배지 보류.",
        "tags": [
          "장소"
        ],
        "operationId": "get__places_id_trend"
      }
    },
...
    "/places/search-assist": {
      "get": {
        "summary": "한국어 이름을 Google 검색으로 수집 장소에 연결",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Place"
                      }
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "MATCHED",
                        "NO_MATCH",
                        "UNAVAILABLE"
                      ]
                    },
                    "notice": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 다른 사용자의 리소스",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100
            }
          },
          {
            "name": "regionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "일반 검색에 결과가 없을 때 사용. 최대 5개 외부 결과를 이름·150m 위치·지역/분류로 기존 PostGIS 장소와 대조한다. 연결할 수 없는 외부 결과를 저장하거나 가상 장소를 생성하지 않는다.",
        "tags": [
          "장소"
        ],
        "operationId": "get__places_search_assist"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kita_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "currentRevision": {
            "type": "integer",
            "minimum": 0,
            "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Region": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ja": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "image": {
            "type": "string"
          }
        }
      },
...
    }
  }
}